added timings to the stats and also working on writing to disk
git-svn-id: https://groke.otago.ac.nz/ant/ant@13 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/memory_index.h
@@ -29,7 +29,7 @@ private:
 	long hash(ANT_string_pair *string);
 	unsigned long dehash(long hash_val);
 	ANT_memory_index_hash_node *find_add_node(ANT_memory_index_hash_node *root, ANT_string_pair *string);
-	long long serialise_all_nodes(ANT_memory_index_hash_node *root, ANT_file *file);
+	long serialise_all_nodes(ANT_memory_index_hash_node *root, ANT_file *file);
 	void text_render(ANT_memory_index_hash_node *root, unsigned char *serialised_docids, long doc_size, unsigned char *serialised_tfs, long tf_size);
 	ANT_memory_index_hash_node *new_memory_index_hash_node(ANT_string_pair *string);
 
@@ -38,7 +38,7 @@ public:
 	~ANT_memory_index();
 
 	void add_term(ANT_string_pair *string, long long docno);
-	long long serialise(char *filename);
+	long serialise(char *filename);
 } ;
